Fixed a download issue for 762 apps
Due to some recent GitHub API updates, it was not possible to download the latest release available from the main page.

The common solution was to modify the download command just enough to check the availability of the direct link from the JSON version of the site page.

PS: I am still checking if there are any download problems for other programs. For the time being 762 scripts have been fixed and should work fine for both AM and AppMan.