Add jetbrains-toolbox
Be careful, this AppImage automatically integrates into the desktop and is difficult to close. It creates files in:
- ~/.config/autostart
 - ~/.local/share/applications

and the option "-H" of AM/AppMan doesn't work with this AppImage.

I don't want to doubt the security of this app, but personally I don't like applications that do what they want on their own and without authorization.

The Appimage may be patched or modified by the authors, please contact https://www.jetbrains.com directly if you have any problems.

See https://github.com/ivan-hc/AM-Application-Manager#spyware-malware-and-dangerous-software to learn more about the reason of the availability of apps like this here.