Prepare the function to notify the user about updating installation scripts
In future updates, it will attempt to notify the user of the availability of a new update script for installed apps. The change will affect the "-s" option (which is included in the "-u" option).

In this commit, the original installation script is copied to a subfolder called ".AM-installer" during program installation ("-i" or "install" option). It will be compared with the one available in the "AM" repository.